+++ src/event-custom/HISTORY.md
@@ -4,6 +4,12 @@ Custom Event Infrastructure Change History
 @VERSION@
 ------
 
+* Fixed regression introduced in 3.10.0, where `EventTarget.detach('cat|*')` 
+  would throw an exception, when the EventTarget was configured with a prefix.
+
+3.11.0
+------
+
 * Fixed issue with fireOnce subscribers not receiving the facade,
   if subscription came in after the fire, and the initial fire had
   no listeners (the bug was introduced in 3.10.0, with the no listener